- Added SMB1 audio playing for death, coin collection, jumping and the background music

+++ POW_SMB1.cpp
@@ -30,7 +30,7 @@ POW::POW(SDL_Renderer* renderer, const char* filePathToSpriteSheet, Vector2D bot
 	}
 	else
 	{
-		mSingleSpriteWidth = mSpriteSheet->GetWidth() / 3.0f;
+		mSingleSpriteWidth = int(mSpriteSheet->GetWidth() / 3.0f);
 	}
 }
